+++ CHANGELOG.md
@@ -4,6 +4,53 @@ All notable changes to this project will be documented in this file. This projec
 adheres to [PEP 440](https://www.python.org/dev/peps/pep-0440/) versioning schema, and
 the changelog itself conforms to [Keep A Changelog](https://keepachangelog.com/).
 
+## v9.11.3 (2026-01-23)
+
+### Fix
+
+-   **updating**: include non-question answers when generating fresh copy of new
+    template
+-   **updating**: ignore Git hooks on internal checkout before 3-way merging file with
+    conflicts (#2432)
+-   avoid pattern deprecation warning for `pathspec` v1.0.0+
+
+## v9.11.2 (2026-01-20)
+
+### Fix
+
+-   **updating**: restore support for preserved symlinks pointing outside subproject
+    (#2427)
+
+### Security
+
+-   disallow symlink-based includes outside template root
+-   disallow symlink-following write operations outside destination directory (#2427)
+
+## v9.11.1 (2026-01-10)
+
+### Fix
+
+-   **updating**: avoid circular reference when rendering JSON-serialized `_copier_conf`
+    variable
+
+## v9.11.0 (2025-11-20)
+
+### Feat
+
+-   **updating**: allow updating a dirty Git repository when the subproject directory is
+    clean (#2369)
+-   add support for custom question icons (#2381)
+-   add support for conditionally unsetting a question's default value
+
+### Fix
+
+-   raise warning instead of error when `chmod` is not allowed
+-   fix using default answers from settings for required questions (#2374)
+
+### Refactor
+
+-   drop support for Python 3.9
+
 ## v9.10.3 (2025-10-17)
 
 ### Fix